pyproject.toml
[project]
name = "clustera-youtube-ingest"
version = "0.1.0"
description = "YouTube data ingestion pipeline for Clustera platform"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "apify-client>=1.0.0",
    "pandas>=1.5.0",
    "aiohttp>=3.10.11",
    "kafka-python>=2.2.10",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]

[project.scripts]
clustera-youtube-ingest = "clustera_youtube_ingest.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
